fix: Cards/Libraries/Relations
removed class-transformer lib
fixed person address cascade
fixed card relations

+++ src/modules/Services/Cards/DeleteCardService.ts
@@ -5,6 +5,7 @@ import Card from '@modules/models/Card/Card';
 
 interface IRequest {
     id: string;
+    user_id: string;
 } 
 
 interface IResponse {
@@ -14,17 +15,22 @@ interface IResponse {
 @injectable()
 class DeleteCardService {
   public async execute({
-    id
+    id,
+    user_id,
   }: IRequest): Promise<IResponse> {
     const cardsRepository = new GenericRepositoryProvider(Card);
     const cardExists = await cardsRepository.findOne({
       where:{
-        id:id,
+        id
       },
     })
 
     if(!cardExists){
         throw new AppError('Cartão não encontrado.')
+    }
+
+    if(cardExists.person.user_id !== user_id){
+        throw new AppError('Usuário não autorizado.')
     }
 
     cardsRepository.remove(cardExists)
